Update SPARKlib project files to remove use of SPARKLIB_OBJECT_DIR
SPARKlib was forcing the definition of a global variable to define
a suitable object file for compiling its sources when used in a client
project. This was not ideal, as use of a local path meant that the object
directory was created directly inside SPARK installation, instead of being
local to the client directories.

Now SPARKlib projects are divided into "external" ones which cannot be used
directly, as they specify Externally_Built to be True, and "model" projects that
a user should copy and adapt to their usage, in terms of Object_Dir and
Excluded_Source_Files. The model ones extends the external ones.

Issue: eng/spark/spark2014#55