Add boolean parameter to handle logical equality in sequences
Add a new boolean parameter Use_Logical_Equality to the functional
sequences generic package. If it is set to True, a lemma has to be
proved on the instance, to ensure that the provided equality is the
logical equality on elements. An automatically instantiated
extensionality axiom is generated for equality on sequences.